+++ tests/tests_cycle_decomposition/test_cases.py
@@ -52,6 +52,13 @@ TEST_DESCENTS = [
     (CycleDecomposition(Cycle(1), Cycle(2), Cycle(3)), []),
     (CycleDecomposition(Cycle(2, 3), Cycle(4, 5, 1)), [1, 2, 4]),
 ]
+TEST_EXCEEDANCES = [
+    (CycleDecomposition(Cycle(1, 2), Cycle(3)), False, [1]),
+    (CycleDecomposition(Cycle(1, 2), Cycle(3)), True, [1, 3]),
+    (CycleDecomposition(Cycle(2, 3), Cycle(4, 5, 1)), False, [1, 2, 4]),
+    (CycleDecomposition(Cycle(1), Cycle(2), Cycle(3)), False, []),
+    (CycleDecomposition(Cycle(1), Cycle(2), Cycle(3)), True, [1, 2, 3]),
+]
 TEST_INVERSE = [
     (CycleDecomposition(Cycle(1, 2, 3)), CycleDecomposition(Cycle(3, 2, 1))),
     (CycleDecomposition(Cycle(1, 2), Cycle(3, 4)), CycleDecomposition(Cycle(2, 1), Cycle(4, 3))),